Handle both `javax.inject.*` and `jakarta.inject.*`.
Previously AutoFactory recognized and generated `javax.inject` only. Now, it will continue to do so if `javax.inject` is present on the classpath, but it will recognize `jakarta.inject` if only that is on the classpath.

Also remove some obsolete tests that would otherwise have had to be updated.

Closes #1184.

RELNOTES=AutoFactory now recognizes and generates both `javax.inject` and `jakarta.inject`. If `jakarta.inject` is on the classpath then it will be used, and otherwise `javax.inject`. A new compiler option `-Acom.google.auto.factory.InjectApi` can be set to either `javax` or `jakarta` to force the use of one or the other.
PiperOrigin-RevId: 535324709